Remove query execution time measurements The move to a transactional based method for executing statements will require a new measurement system. We'll be measuring transaction execution time and won't have access to every sub-query execution time.
This functionality is probably better handled through use of the EXPLAIN
statement where much more detailed and useful data can be captured. This
is not defined within the SQL standard and will need to be considered
seperately for each database server type. Arguably MantisBT shouldn't be
getting involved with this level of debugging -- it is probably better
left to the tools and features provided by the database server which are
far more powerful and better supported.